+++ lib/forms/Button/Button.tsx
@@ -1,6 +1,6 @@
 import type React from "react";
 import { Button as BaseButton } from "@/components/ui/button";
-import { Flex } from "@/primitives";
+import { Flex } from "@/layout";
 
 export type BaseButtonVariant = "default" | "secondary" | "destructive" | "outline" | "ghost" | "link";
 export type BaseButtonSize = "default" | "sm" | "lg" | "icon";
