+++ lib/layout/index.ts
@@ -1,5 +1,3 @@
 export * from "./Box";
 export * from "./Flex";
 export * from "./Grid";
-export * from "./Heading";
-export * from "./Text";
