+++ lib/main.ts
@@ -1,22 +1,3 @@
-export type { ButtonGroupProps, ButtonProps } from "@/forms/Button";
-export { Button, ButtonGroup, DeleteButton, SaveButton } from "@/forms/Button";
-export type { LabelProps } from "@/forms/Label";
-export { Label } from "@/forms/Label";
-export type { NumberInputProps } from "@/forms/NumberInput";
-export { NumberInput } from "@/forms/NumberInput";
-export type { SelectPickerOption, SelectPickerProps } from "@/forms/SelectPicker";
-export { SelectPicker } from "@/forms/SelectPicker";
-export type { TextareaProps } from "@/forms/Textarea";
-export { Textarea } from "@/forms/Textarea";
-export type { TextInputProps } from "@/forms/TextInput";
-export { TextInput } from "@/forms/TextInput";
-export type { BoxProps } from "@/primitives/Box";
-export { Box } from "@/primitives/Box";
-export type { FlexProps } from "@/primitives/Flex";
-export { Flex } from "@/primitives/Flex";
-export type { GridProps } from "@/primitives/Grid";
-export { Grid } from "@/primitives/Grid";
-export type { HeadingProps } from "@/primitives/Heading";
-export { Heading } from "@/primitives/Heading";
-export type { TextProps } from "@/primitives/Text";
-export { Text } from "@/primitives/Text";
+export * from "./forms";
+export * from "./layout";
+export * from "./typography";
