+++ lib/typography/index.ts
@@ -1,0 +1,2 @@
+export * from "./Heading";
+export * from "./Text";
