+++ stories/templates/OptionsByFamilyGrid.tsx
@@ -1,7 +1,8 @@
 import Code from "@src/components/Code";
 import { Fragment } from "react";
-import { Box, Flex, Grid, type HeadingProps, Text } from "@/primitives";
+import { Box, Flex, Grid } from "@/layout";
 import { FONT_FAMILIES, type FontFamily } from "@/types";
+import { HeadingProps, Text } from "@/typography";
 
 type ValueOf<T> = T[keyof T];
 type OptionTypes = ValueOf<HeadingProps>;
